requirements.txt
ansi2html==1.8.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
drawnow==0.72.5
Flask==2.2.5
fonttools==4.40.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.15.0
pyparsing==3.1.0
python-can==4.2.2
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==8.2.2
typing_extensions==4.7.0
tzdata==2023.3
uptime==3.0.1
urllib3==2.0.3
Werkzeug==2.2.3
wrapt==1.15.0